ConsolidateStaticCacheUpdates: merge inference into the pass itself
Per reviewer feedback: the previous design split analysis (mark args with
the ttcore.cumulative_length attribute, in a separate
TTIRInferStaticCacheCumulativeLengthArgs pass) from transformation (the
consolidation pass that consumed the attribute), modeled on
TTIRInferKVCacheArgumentTypes. But kv_cache has multiple downstream
consumers (TTNNKVCacheDtypeConversion, TTNNLayout, RowMajorLayoutPropagation,
the Utils.h hasKVCacheArgAttr helper), so the split there earns its keep.

cumulative_length had exactly one consumer — the consolidation pass — so
the separation was structure without payoff.

Changes:
- Move the back-tracing logic (collect entry-block args feeding an
  update-style cache op's update_index) into a static helper in
  ConsolidateStaticCacheUpdates.cpp.
- runOnOperation builds the per-function set of cumulative_length args
  inline, then gates the candidate matcher on membership in that set.
- Delete InferStaticCacheCumulativeLengthArgs.cpp,
  TTIRInferStaticCacheCumulativeLengthArgs Passes.td entry, the
  g_cumulativeLengthAttrName constant in Utils.h, the inference-pass entry
  in the TTNN pipeline, and the CMakeLists.txt entry.
- Update the consolidate lit test to anchor cumulative_length args via real
  `ttir.update_cache` ops (chained when there are multiple), replacing the
  previous `{ttcore.cumulative_length}` arg attribute annotations. The new
  negative case `no_consolidate_no_cache_op` verifies that without a cache
  op, the pass is a no-op (the property previously verified by
  `no_consolidate_unmarked_args`).
- Delete the standalone inference-pass lit test; its coverage is now folded
  into the consolidate test via the chained update_cache cases.

Semantic guarantee is unchanged: an arg is only eligible for unification if
it provably feeds a cache update_index, exactly as before. The proof is now
internal to the pass instead of materialized as an IR-level attribute.
