+++ include/ttmlir/Dialect/TTCore/IR/Utils.h
@@ -18,8 +18,6 @@
 namespace mlir::tt::ttcore {
 
 constexpr inline llvm::StringLiteral g_kvCacheAttrName = "ttcore.kv_cache";
-constexpr inline llvm::StringLiteral g_cumulativeLengthAttrName =
-    "ttcore.cumulative_length";
 
 constexpr inline llvm::StringLiteral g_originalActivationNamesAttrName =
     "ttcore.original_activation_names";
