+++ lib/Dialect/TTIR/Transforms/CMakeLists.txt
@@ -14,7 +14,6 @@ add_mlir_dialect_library(MLIRTTIRTransforms
         ElementTypeNormalization.cpp
         PredicateTypeAlignment.cpp
         InferKVCacheArgumentTypes.cpp
-        InferStaticCacheCumulativeLengthArgs.cpp
         PropagateWeightDtype.cpp
         MoveReshapeToConstant.cpp
         Quantization.cpp
