+++ lib/Dialect/TTNN/Pipelines/TTNNPipelines.cpp
@@ -92,8 +92,6 @@ void createTTNNPipelineTTIRPasses(
     pm.addPass(mlir::tt::ttir::createTTIRFusing(fusingOptions));
   }
   pm.addPass(mlir::tt::ttir::createTTIRFoldFullToScalar());
-  pm.addPass(
-      mlir::tt::ttir::createTTIRInferStaticCacheCumulativeLengthArgs());
   pm.addPass(mlir::tt::ttir::createTTIRConsolidateStaticCacheUpdates());
   pm.addPass(mlir::createCSEPass());
 }
